Proyecto1/requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter Environment
jupyter>=1.0.0
ipykernel>=6.0.0

# Presentation Generation
python-pptx>=0.6.0

# Additional Utilities
joblib>=1.2.0
warnings
json
os
time
itertools

# Optional: Advanced ML Libraries (comment out if not needed)
# xgboost>=1.6.0
# lightgbm>=3.3.0 